+++ scripts/install-kind.sh
@@ -1,0 +1,5 @@
+#! /bin/bash
+
+curl -Lo ./kind https://kind.sigs.k8s.io/dl/v0.11.1/kind-linux-amd64
+chmod +x ./kind
+mv ./kind /usr/local/bin/kind